refactor(repositories): rename course_content.py -> course_content_queries.py
repositories/course_content.py (1275 LOC) has NO repository class -- it is a
collection of ~15 free SQLAlchemy query/subquery builders + one DTO. The name
made it look like the entity repository, which actually lives in
course_content_repo.py (CourseContentRepository) -- a real navigation trap.

Rename the module to course_content_queries.py and update its docstring to say
what it is. Updated the 7 import sites across the 5 importers (student_view,
tutor_view, view_base, view_mappers, business_logic/tutor); git tracks the file
as a rename. No code logic changed.

This is the de-misnaming step; a by-concern split (results / submissions /
messages subqueries -- which have zero external importers -- vs. composite
content queries vs. per-member count aggregates) is a natural follow-up.

Verified: renamed module + all importers byte-compile and import.
